Use junit's ExpectedException in JobsTest
Use ExpectedException instead of the exception argument in the @Test
annotation. Change the test to be more specific about what exception
is expected, and also add a test for the expected error message.

The ExpectedException is declared in the base class so it can also be
used in other tests later.

Change-Id: Ibec9ac1e110752c5cbfa3c95459da485e8dfee4e